+++ tests/Phan/Language/EmptyUnionTypeTest.php
@@ -211,11 +211,17 @@ final class EmptyUnionTypeTest extends BaseTest
                 ];
             case Closure::class:
                 return [
-                    /** @param mixed ...$unused_args */
+                    /**
+                     * @param mixed ...$unused_args
+                     * @suppress PhanPluginUnknownArrayClosureParamType
+                     */
                     static function (...$unused_args): bool {
                         return false;
                     },
-                    /** @param mixed ...$unused_args */
+                    /**
+                     * @param mixed ...$unused_args
+                     * @suppress PhanPluginUnknownArrayClosureParamType
+                     */
                     static function (...$unused_args): bool {
                         return true;
                     },